+++ src/rgw/rgw_rest_s3.cc
@@ -1186,11 +1186,12 @@ int RGWPostObj_ObjStore_S3::get_policy()
         return -EACCES;
       }
 
-      user_info.user_id = keystone_validator.response.token.tenant.id;
-      user_info.display_name = keystone_validator.response.token.tenant.name;
+      string project_id = keystone_validator.response.get_project_id();
+      user_info.user_id = project_id;
+      user_info.display_name = keystone_validator.response.get_project_name();
 
       /* try to store user if it not already exists */
-      if (rgw_get_user_info_by_uid(store, keystone_validator.response.token.tenant.id, user_info) < 0) {
+      if (rgw_get_user_info_by_uid(store, project_id, user_info) < 0) {
         int ret = rgw_store_user_info(store, user_info, NULL, NULL, 0, true);
         if (ret < 0) {
           dout(10) << "NOTICE: failed to store new user's info: ret=" << ret << dendl;
@@ -2243,9 +2244,15 @@ int RGWHandler_ObjStore_S3::init(RGWRados *store, struct req_state *s, RGWClient
 int RGW_Auth_S3_Keystone_ValidateToken::validate_s3token(const string& auth_id, const string& auth_token, const string& auth_sign) {
   /* prepare keystone url */
   string keystone_url = cct->_conf->rgw_keystone_url;
+  string keystone_version = cct->_conf->rgw_keystone_api_version;
   if (keystone_url[keystone_url.size() - 1] != '/')
     keystone_url.append("/");
-  keystone_url.append("v2.0/s3tokens");
+  if (keystone_version == "3") {
+    keystone_url.append("v3/s3tokens");
+  }
+  else {
+    keystone_url.append("v2.0/s3tokens");
+  }
 
   /* set required headers for keystone request */
   append_header("X-Auth-Token", cct->_conf->rgw_keystone_admin_token);
@@ -2261,11 +2268,11 @@ int RGW_Auth_S3_Keystone_ValidateToken::validate_s3token(const string& auth_id, 
   /* create json credentials request body */
   JSONFormatter credentials(false);
   credentials.open_object_section("");
-  credentials.open_object_section("credentials");
-  credentials.dump_string("access", auth_id);
-  credentials.dump_string("token", token_encoded.c_str());
-  credentials.dump_string("signature", auth_sign);
-  credentials.close_section();
+    credentials.open_object_section("credentials");
+      credentials.dump_string("access", auth_id);
+      credentials.dump_string("token", token_encoded.c_str());
+      credentials.dump_string("signature", auth_sign);
+    credentials.close_section();
   credentials.close_section();
 
   std::stringstream os;
@@ -2289,7 +2296,7 @@ int RGW_Auth_S3_Keystone_ValidateToken::validate_s3token(const string& auth_id, 
   bool found = false;
   list<string>::iterator iter;
   for (iter = roles_list.begin(); iter != roles_list.end(); ++iter) {
-    if ((found=response.user.has_role(*iter))==true)
+    if ((found=response.has_role(*iter))==true)
       break;
   }
 
@@ -2299,7 +2306,7 @@ int RGW_Auth_S3_Keystone_ValidateToken::validate_s3token(const string& auth_id, 
   }
 
   /* everything seems fine, continue with this user */
-  ldout(cct, 5) << "s3 keystone: validated token: " << response.token.tenant.name << ":" << response.user.name << " expires: " << response.token.expires << dendl;
+  ldout(cct, 5) << "s3 keystone: validated token: " << response.get_project_name() << ":" << response.get_user_name() << " expires: " << response.get_expires() << dendl;
   return 0;
 }
 
@@ -2387,11 +2394,12 @@ int RGW_Auth_S3::authorize(RGWRados *store, struct req_state *s)
 	}
 
 
-	s->user.user_id = keystone_validator.response.token.tenant.id;
-        s->user.display_name = keystone_validator.response.token.tenant.name; // wow.
+        string project_id = keystone_validator.response.get_project_id();
+	      s->user.user_id = project_id;
+        s->user.display_name = keystone_validator.response.get_project_name(); // wow.
 
         /* try to store user if it not already exists */
-        if (rgw_get_user_info_by_uid(store, keystone_validator.response.token.tenant.id, s->user) < 0) {
+        if (rgw_get_user_info_by_uid(store, project_id, s->user) < 0) {
           int ret = rgw_store_user_info(store, s->user, NULL, NULL, 0, true);
           if (ret < 0)
             dout(10) << "NOTICE: failed to store new user's info: ret=" << ret << dendl;
