+++ src/rgw/rgw_rest_s3.h
@@ -308,8 +308,9 @@ private:
   }
 
 public:
-  RGW_Auth_S3_Keystone_ValidateToken(CephContext *_cct)
-      : RGWHTTPClient(_cct) {
+  RGW_Auth_S3_Keystone_ValidateToken(CephContext *_cct) :
+          RGWHTTPClient(_cct),
+          response(KeystoneToken(_cct->_conf->rgw_keystone_api_version)) {
     get_str_list(cct->_conf->rgw_keystone_accepted_roles, roles_list);
   }
